Implemented ostream and istream
I implemented ostream and istream across the main.cpp where possible. I also had to make changes to the original ostream and istream operators from my previous commit because they weren't working as planned.

Note that the commit shows that I completely changed each file, this is because I was working on a separate copy of the project and copy and pasted my work over. I did this to compare functionality with the original code so that I could achieve the same functionality with ostream and istream as we previously did.